extract some logic from fly to shared
