+++ waspc/packages/deploy/src/providers/fly/cmd/cmd.ts
@@ -1,7 +1,8 @@
 import { exit } from 'process';
 import { $, cd } from 'zx';
+import { getCommandHelp, waspSays } from '../../shared/helpers.js';
 import { CommonOps, getCommonOps } from '../helpers/CommonOps.js';
-import { buildDirExists, getCommandHelp, waspSays } from '../helpers/helpers.js';
+import { buildDirExists } from '../helpers/helpers.js';
 import { deleteLocalToml, getTomlFilePaths, localTomlExists } from '../helpers/tomlFileHelpers.js';
 import { executeFlyCommand } from '../index.js';
 import { CmdOptions } from './CmdOptions.js';
