+++ waspc/packages/deploy/src/providers/fly/deploy/deploy.ts
@@ -3,10 +3,6 @@ import { $, cd, fs } from 'zx';
 import {
 	cdToClientBuildDir,
 	cdToServerBuildDir,
-	displayWaspRocketImage,
-	makeIdempotent,
-	getCommandHelp,
-	waspSays,
 } from '../helpers/helpers.js';
 import {
 	clientTomlExistsInProject,
@@ -23,6 +19,7 @@ import { DeployOptions } from './DeployOptions.js';
 import { createDeploymentInfo, DeploymentInfo } from '../DeploymentInfo.js';
 import { flySetupCommand } from '../index.js';
 import { secretExists } from '../helpers/flyctlHelpers.js';
+import { displayWaspRocketImage, getCommandHelp, makeIdempotent, waspSays } from '../../shared/helpers.js';
 
 export async function deploy(options: DeployOptions): Promise<void> {
 	waspSays('Deploying your Wasp app to Fly.io!');
