+++ waspc/packages/deploy/src/providers/fly/helpers/flyctlHelpers.ts
@@ -1,7 +1,7 @@
 import { exit } from 'process';
 import { $, question } from 'zx';
 import { executeFlyCommand } from '../index.js';
-import { silence, isYes, waspSays, getCommandHelp } from './helpers.js';
+import { silence, waspSays, getCommandHelp, isYes } from '../../shared/helpers.js';
 
 export async function flyctlExists(): Promise<boolean> {
 	try {
