+++ waspc/packages/deploy/src/providers/fly/helpers/helpers.ts
@@ -1,12 +1,9 @@
 import { Command } from 'commander';
 import { exit } from 'process';
-import { $, cd, ProcessOutput, Shell } from 'zx';
+import { cd } from 'zx';
 import fs from 'fs';
 import path from 'node:path';
-
-export function isYes(str: string): boolean {
-	return str.trim().toLowerCase().startsWith('y');
-}
+import { waspSays } from '../../shared/helpers.js';
 
 export function ensureWaspDirLooksRight(thisCommand: Command): void {
 	const dirContainsWasproot = fs.existsSync(path.join(thisCommand.opts().waspProjectDir, '.wasproot'));
@@ -66,62 +63,4 @@ export function ensureDirsInCmdAreAbsoluteAndPresent(thisCommand: Command): void
 			exit(1);
 		}
 	}
-}
-
-// eslint-disable-next-line
-export function makeIdempotent<F extends () => any>(
-	fn: F,
-): () => ReturnType<F> {
-	let result: { value: ReturnType<F> } | null = null;
-
-	return function idempotentFn() {
-		if (!result) {
-			result = { value: fn() };
-		}
-		return result.value;
-	};
-}
-
-// For some reason, the colors from the chalk package wouldn't
-// show up when run as a subprocess by the Wasp CLI. This works.
-export function waspSays(str: string): void {
-	console.log('🚀 \x1b[33m ' + str + ' \x1b[0m');
-}
-
-export function displayWaspRocketImage(): void {
-	// Escaping backslashes makes it look weird here, but it works in console.
-	const asciiArt = `
-
-                    __
-                   // \\
-                   \\\\_/ //
-             _    -(||)(')
-            \\ \\____///_____
-   #########[==__DEPLOYED__}
-            /_/
-
-  `;
-	console.log(asciiArt);
-}
-
-export function getCommandHelp(command: Command): string {
-	return trimUsage(command.helpInformation());
-}
-
-function trimUsage(usage: string): string {
-	return usage.split(/[\r\n]+/)[0].replace('Usage: ', '').replace(' [options]', '');
-}
-
-// There is a theoretical race condition here since we are modifying a global `$`
-// property, that when we yield to the `await cmd($)` call that some other calls to
-// `$` could use a different verbosity setting. Additionally, calling `silence` multiple
-// times concurrently could change the setting incorrectly.
-// However, our pattern of awaiting for both `$` and `silence` calls without any random
-// callbacks using either means this interleaving should not ever happen.
-export async function silence(cmd: ($hh: Shell) => Promise<ProcessOutput>): Promise<ProcessOutput> {
-	const verboseSetting = $.verbose;
-	$.verbose = false;
-	const proc = await cmd($);
-	$.verbose = verboseSetting;
-	return proc;
 }
