+++ waspc/packages/deploy/src/providers/fly/setup/setup.ts
@@ -11,8 +11,9 @@ import {
 } from '../helpers/tomlFileHelpers.js';
 import { createDeploymentInfo, DeploymentInfo } from '../DeploymentInfo.js';
 import { CommonOptions } from '../CommonOptions.js';
-import { cdToClientBuildDir, cdToServerBuildDir, makeIdempotent, getCommandHelp, waspSays } from '../helpers/helpers.js';
+import { cdToClientBuildDir, cdToServerBuildDir } from '../helpers/helpers.js';
 import { createFlyDbCommand } from '../index.js';
+import { getCommandHelp, makeIdempotent, waspSays } from '../../shared/helpers.js';
 
 export async function setup(baseName: string, region: string, options: CommonOptions): Promise<void> {
 	waspSays('Setting up your Wasp app with Fly.io!');
